Makes a method module; all tests and spec harness still passing for its 1 and 0
