+++ lib/item_repo.rb
@@ -1,26 +1,16 @@
-# require 'CSV'
 require_relative '../lib/item'
 require_relative '../lib/sales_engine'
+require_relative '../lib/method_module'
 
 class ItemRepo
+  include FindMethods
+
   attr_reader :all
 
   def initialize(item_data)
     @all = item_data
     @first_step = first_step
     @sum = sum
-  end
-
-  def find_by_id(id)
-    @all.find do |item|
-      item.id == id
-    end
-  end
-
-  def find_by_name(name)
-    @all.find do |item|
-      item.name.upcase == name.upcase
-    end
   end
 
   def find_all_with_description(description)
@@ -47,13 +37,6 @@ class ItemRepo
     end
   end
 
-  def find_highest_id
-     current_highest = @all.max_by do |item|
-       item.id
-     end
-     current_highest.id
-  end
-
   def create(attributes)
    new_id = find_highest_id + 1
    attributes[:id] = new_id
@@ -70,12 +53,6 @@ class ItemRepo
        single_item.updated_at = Time.now
      end
      single_item
-  end
-
-  def delete(id)
-    @all.delete_if do |item|
-      item.id == id
-    end
   end
 
   def average_item_price
